move chat log event handler into its own file
